- SF# 3559325: Util Windows broken in non-Unicode - port of rev. 1918 from 1.4.4 branch (fixed SF# 3519474) - update VS 2010 Util projects with JSON

+++ Util/testsuite/src/WinRegistryTest.cpp
@@ -55,11 +55,9 @@ WinRegistryTest::~WinRegistryTest()
 void WinRegistryTest::testRegistry()
 {
 	WinRegistryKey regKey("HKEY_CURRENT_USER\\Software\\Applied Informatics\\Test");
-	if (regKey.exists())
-	{
-		regKey.deleteKey();
-	}
+	if (regKey.exists()) regKey.deleteKey();
 	assert (!regKey.exists());
+
 	regKey.setString("name1", "value1");
 	assert (regKey.getString("name1") == "value1");
 	regKey.setString("name1", "Value1");
